requirements.txt
cloudinary==1.44.1
dj3-cloudinary-storage==0.0.6
Django==5.2.6
Faker==37.8.0
gunicorn==23.0.0
mysqlclient==2.2.7
pillow==11.3.0
psycopg2-binary==2.9.10
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1